refactor: move html, css and assets to webdev-infra
