+++ shortcodes/BrowserCompat/styles.scss
@@ -1,0 +1,75 @@
+.wdi-browser-compat {
+  align-items: center;
+  color: var(--wdi-text-color, #585b63);
+  display: flex;
+  flex-wrap: wrap;
+
+  @include media-query('md') {
+    flex-wrap: nowrap;
+  }
+
+  &__icon {
+    background: content-box no-repeat transparent;
+    display: inline-block;
+    height: 24px;
+    margin-left: 24px;
+    margin-right: 6px;
+    width: 24px;
+  }
+
+  &__items {
+    align-items: center;
+    display: flex;
+    margin: 1rem 1rem 1rem 0;
+  }
+
+  &__version {
+    --size-2: clamp(0.75rem, 0.71rem + 0.18vw, 0.875rem); // From web.dev
+    border-radius: 100%;
+    display: inline-block;
+    font-size: var(--size-2);
+    height: 20px;
+    line-height: 20px;
+    min-width: 20px;
+    text-align: center;
+  
+    &[data-compat='yes'] {
+      background: var(--wdi-success-bg-color, #e9f6ed);
+      color: var(--wdi-success-color, #0d652d);
+    }
+  
+    &[data-compat='preview'], &[data-compat='flag'] {
+      background: var(--wdi-warn-bg-color, #fff5e3);
+      color: var(--wdi-warn-color, #c34900);
+    }
+  
+    &[data-compat='deprecated'], &[data-compat='no'] {
+      background: var(--wdi-error-bg-color, #fce8e8);
+      color: var(--wdi-error-color, #a50e0e);
+    }
+  }
+
+  &__label {
+    color: var(--wdi-text-color, #585b63);
+    flex: 0 0 100%;
+    font-style: normal;
+    margin-right: 1rem;
+    width: 100%;
+  
+    @include media-query('md') {
+      flex: none;
+      margin-bottom: 0;
+      width: auto;
+    }
+  }
+  
+  &__link {
+    flex: 0 0 100%;
+    margin-left: 0;
+  
+    @include media-query('md') {
+      flex: none;
+      margin-left: 24px;
+    }
+  }
+}
